+++ python/pathway/io/airbyte/__init__.py
@@ -17,7 +17,6 @@ from pathway.third_party.airbyte_serverless.executable_runner import (
     AbstractAirbyteSource,
 )
 
-INCREMENTAL_SYNC_MODE = "incremental"
 METHOD_PYPI = "pypi"
 METHOD_DOCKER = "docker"
 
@@ -49,6 +48,10 @@ def _construct_local_source(
     enforce_method: str | None = None,
 ) -> AbstractAirbyteSource:
     with optional_imports("airbyte"):
+        from pathway.io.airbyte.logic import (
+            FULL_REFRESH_SYNC_MODE,
+            INCREMENTAL_SYNC_MODE,
+        )
         from pathway.third_party.airbyte_serverless.sources import (
             DockerAirbyteSource,
             VenvAirbyteSource,
@@ -81,11 +84,18 @@ def _construct_local_source(
         )
 
     # Run airbyte connector locally and check streams
+    global_sync_mode = None
     for stream in source.configured_catalog["streams"]:
         name = stream["stream"]["name"]
         sync_mode = stream["sync_mode"]
-        if sync_mode != INCREMENTAL_SYNC_MODE:
-            raise ValueError(f"Stream {name} doesn't support 'incremental' sync mode")
+        if sync_mode != INCREMENTAL_SYNC_MODE and sync_mode != FULL_REFRESH_SYNC_MODE:
+            raise ValueError(f"Stream {name} has unknown sync_mode: {sync_mode}")
+        global_sync_mode = global_sync_mode or sync_mode
+        if global_sync_mode != sync_mode:
+            raise ValueError(
+                "All streams within the same 'pw.io.airbyte.read' must have "
+                "the same 'sync_mode'"
+            )
 
     return source
 
@@ -304,6 +314,7 @@ mode.
         )
         source = RemoteAirbyteSource(
             config=config,
+            streams=streams,
             job_id=job_id,
             credentials=credentials,
             region=gcp_region,