+++ python/pathway/tests/utils.py
@@ -37,7 +37,7 @@ xfail_on_multiple_threads = pytest.mark.xfail(
     os.getenv("PATHWAY_THREADS", "1") != "1", reason="multiple threads"
 )
 
-AIRBYTE_CONNECTION_REL_PATH = "connections/new_source.yaml"
+AIRBYTE_FAKER_CONNECTION_REL_PATH = "connections/faker.yaml"
 
 
 def skip_on_multiple_workers() -> None: