+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - **Experimental** A ``pw.xpacks.llm.document_store.DocumentStore`` to process and index documents.
 - ``pw.xpacks.llm.servers.DocumentStoreServer`` used to expose REST server for retrieving documents from ``pw.xpacks.llm.document_store.DocumentStore``.
 - `pw.xpacks.stdlib.indexing.HybridIndex` used for querying multiple indices and combining their results.
+- `pw.io.airbyte.read` now also supports streams that only operate in `full_refresh` mode.
 
 ### Changed
 - Running servers for answering queries is extracted from `pw.xpacks.llm.question_answering.BaseRAGQuestionAnswerer` into `pw.xpacks.llm.servers.QARestServer` and `pw.xpacks.llm.servers.QASummaryRestServer`.